Demostrate problem with nonchemical elements.
We can't build storage in the XYZTranslator due to the use of nonchemical
(arbitrary) elements. I suspect the problem is that mdtraj's
Topology.from_openmm() doesn't load changes to the element database in the
openmm.app.element class. (The test fails with a key error when
mdtraj.element.Element._elements_by_symbol[key] faces a key from the newly
added elements.) Committing this so others can see what's happening and
advise on solutions.