Skeletons of OrgByEns and tests. Needs some redo.
In particular, there's a problem that the make_choosers was only working if
there was actually a mover of each type for each ensemble. Of course, this
isn't the case: for the minus ensemble, there's only one legal type of move!

Still need to do some serious cleanup to get this working.