Removed old testSeqEns.test_minus*.
There was an error in the so-called "minus" ensemble used there, so the
results it would give did not match what we'd expect from the real minus
ensemble for can_prepend. The real minus ensemble is thoroughly tested in
another test class, so that's what we should be using. Namespace clashes
(two tests with same name) were causing us to not see that one of the tests
was wrong.

This also includes some starts on the proper strict_can_* for SeqEns.