Fix to make MSOuter.make_ensemble match docstring
The docstring said that MSOuterEnsemble.make_ensemlbe takes a
parameter `forbidden`, which is a list of `Volume`s. However, it
only worked if `forbidden` was a single `Volume`. This fixes it so
that it can take either a single `Volume` or a list of them.

This popped up as a problem when doing a MISTISNetwork with
`strict_sampling=True`. The MISTISNetwork code expected MSOuter
to do what the docstring said.