Update fixtures
This links the make_tis_trajectory method to a fixture class that
depends on the details of the setup. In this way, there's less of
an implicit dependence on a certain kind of setup. Also switches
it so that the input values for trajectories here are based on
integers instead of floats, which should avoid confusion.

Probably still need significant docs here, but the tests seem to
pass locally, so that's good.