Fail fast if `prepare` fails in external engines
The `prepare` command in external engines can be used to, e.g,
create input for a given trajectory. For example, in Gromacs,
this runs the grompp command to incorporate snapshot-independent
information in the mdp with the specific trr snapshot.

Previously, there was no catch for an error in `prepare`. This
meant that the engine would try to run the trajetory, and, not
finding required files, would fail with the unhelpful message
"engine died unexpectedly." This PR provides a clearer error
message in the case that the `prepare` step fails.

It provides a specific error for the Gromacs engine, and a
generic error for any engine that returns a non-zero result code
from `prepare`, but which hasn't already raised the error.