Storage Handlers
These are intended to abstract out filesystem storage to be an arbitrary
key-value store. As examples, I've implemented storage on the filesystem
and in memory. Object storage (e.g., S3) can be implemented separately.

I'm making this a separate PR so that it can be used in future PRs.

I'd also appreciate suggestions on a better name! `StorageHandler` is
kind of awful.